AudioDeployTool/requirements.txt
questionary>=2.1.1
rich==13.9.4
PyYAML==6.0.2
